project8-agent-system/requirements.txt
crewai>=0.130.0
anthropic>=0.78.0
python-dotenv>=1.0.1
requests>=2.32.3
feedparser>=6.0.11
gspread>=6.1.2
google-auth>=2.38.0
pytest>=8.3.5
bandit>=1.8.3
semgrep>=1.120.0
pyyaml>=6.0.2
